Decouple changing of list-item depth from event handler
