+++ packages/elements/list/src/transforms/index.ts
@@ -2,6 +2,7 @@
  * @file Automatically generated by barrelsby.
  */
 
+export * from './changeListItemDepth';
 export * from './insertListItem';
 export * from './moveListItemDown';
 export * from './moveListItemSublistItemsToListItemSublist';
